Crossposting refactor and code splitting
