+++ packages/lesswrong/server/fmCrosspost/tokens.ts
@@ -1,0 +1,45 @@
+import { DatabaseServerSetting } from "../databaseSettings";
+import jwt, { VerifyErrors } from "jsonwebtoken";
+import { MissingSecretError } from "./errors";
+
+const crosspostSigningKeySetting = new DatabaseServerSetting<string|null>("fmCrosspostSigningKey", null);
+
+const getSecret = () => {
+  const secret = crosspostSigningKeySetting.get();
+  if (!secret) {
+    throw new MissingSecretError();
+  }
+  return secret;
+}
+const jwtSigningOptions = {
+  algorithm: "HS256",
+  expiresIn: "30m",
+} as const;
+
+export const signToken = <T extends {}>(payload: T): Promise<string> =>
+  new Promise((resolve, reject) => {
+    jwt.sign(payload, getSecret(), jwtSigningOptions, (err, token) => {
+      if (token) {
+        resolve(token);
+      } else {
+        reject(err);
+      }
+    });
+  });
+
+export const verifyToken = <T extends {}>(token: string, validator: (payload: T) => void): Promise<T> =>
+  new Promise((resolve, reject) => {
+    jwt.verify(token, getSecret(), (err: VerifyErrors | null, decoded?: T) => {
+      const payload = decoded as T;
+      if (payload) {
+        try {
+          validator(payload);
+        } catch (e) {
+          reject(e);
+        }
+        resolve(payload);
+      } else {
+        reject(err);
+      }
+    });
+  });
